requirements.txt
streamlit>=1.30.0
ollama>=0.4.0
pytest>=8.0.0
plotly>=5.18.0
folium>=0.20.0
streamlit-folium>=0.27.0
mercantile>=1.2.0
google-genai>=1.0.0
python-dotenv>=1.0.0
